# models/model_like_response.go
/*
 * Sample project for KubeQuest
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type LikeResponse struct {

	Dog int64 `json:"dog"`

	Cat int64 `json:"cat"`
}