# models/model_response.go
/*
 * Sample project for KubeQuest
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type Response struct {

	Token string `json:"token"`

	Message string `json:"message"`
}